+++ r2-netty/src/main/java/com/linkedin/r2/netty/client/HttpNettyClient.java
@@ -301,7 +301,6 @@ public class HttpNettyClient implements TransportClient
     try
     {
       TimingContextUtil.markTiming(requestContext, TIMING_KEY);
-      // TODO : convert this TCP Socket address to DomainSocketAddress
       address = resolveAddress(request, requestContext);
       TimingContextUtil.markTiming(requestContext, TIMING_KEY);
     }