+++ r2-netty/src/main/java/com/linkedin/r2/transport/http/client/TransportProtocol.java
@@ -1,6 +1,0 @@
-package com.linkedin.r2.transport.http.client;
-
-public enum TransportProtocol {
-  TCP,
-  UDS
-}