+++ r2-netty/src/main/java/com/linkedin/r2/transport/http/client/common/ChannelPoolManagerFactoryImpl.java
@@ -203,7 +203,7 @@ public class ChannelPoolManagerFactoryImpl implements ChannelPoolManagerFactory
           _enableSSLSessionResumption,
           _connectTimeout,
           _sslHandShakeTimeout,
-          channelPoolManagerKey.getTransportProtocol());
+          channelPoolManagerKey.getUdsAddress());
     }
     else
     {