+++ r2-netty/src/main/java/com/linkedin/r2/transport/http/client/common/ChannelPoolManagerKeyBuilder.java
@@ -18,7 +18,6 @@ package com.linkedin.r2.transport.http.client.common;
 
 import com.linkedin.r2.transport.http.client.AsyncPoolImpl;
 import com.linkedin.r2.transport.http.client.HttpClientFactory;
-import com.linkedin.r2.transport.http.client.TransportProtocol;
 import io.netty.util.internal.ObjectUtil;
 
 import javax.net.ssl.SSLContext;
@@ -47,7 +46,7 @@ public class ChannelPoolManagerKeyBuilder
   private AsyncPoolImpl.Strategy _strategy = HttpClientFactory.DEFAULT_POOL_STRATEGY;
   private boolean _tcpNoDelay = HttpClientFactory.DEFAULT_TCP_NO_DELAY;
   private String _poolStatsNamePrefix = HttpClientFactory.DEFAULT_POOL_STATS_NAME_PREFIX;
-  private TransportProtocol _transportProtocol = HttpClientFactory.DEFAULT_TRANSPORT_PROTOCOL;
+  private String _udsAddress = null;
 
   /**
    * @param sslContext {@link SSLContext}
@@ -201,8 +200,8 @@ public class ChannelPoolManagerKeyBuilder
     return this;
   }
 
-  public ChannelPoolManagerKeyBuilder setTransportProtocol(TransportProtocol transportProtocol) {
-    _transportProtocol = transportProtocol;
+  public ChannelPoolManagerKeyBuilder setUdsAddress(String udsAddress) {
+    _udsAddress = udsAddress;
     return this;
   }
 
@@ -210,6 +209,6 @@ public class ChannelPoolManagerKeyBuilder
   {
     return new ChannelPoolManagerKey(_sslContext, _sslParameters, _gracefulShutdownTimeout, _idleTimeout, _sslIdleTimeout,
       _maxHeaderSize, _maxChunkSize, _maxResponseSize, _maxPoolSize, _minPoolSize, _maxConcurrentConnectionInitializations,
-      _poolWaiterSize, _strategy, _tcpNoDelay, _poolStatsNamePrefix, _transportProtocol);
+      _poolWaiterSize, _strategy, _tcpNoDelay, _poolStatsNamePrefix, _udsAddress);
   }
 }