+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and what APIs have changed, if applicable.
 
 ## [Unreleased]
 
+## [29.37.19] - 2022-09-01
+- Add support of Unix Domain Socket (UDS) Transport Protocol for R2/Netty Client for Outbound transports
+
 ## [29.37.18] - 2022-08-29
 - Support supplying D2 subsetting peer cluster name at run-time
 - 
@@ -5319,6 +5322,7 @@ patch operations can re-use these classes for generating patch messages.
 
 ## [0.14.1]
 
+<<<<<<< HEAD
 [Unreleased]: https://github.com/linkedin/rest.li/compare/v29.37.18...master
 [29.37.18]: https://github.com/linkedin/rest.li/compare/v29.37.17...v29.37.18
 [29.37.17]: https://github.com/linkedin/rest.li/compare/v29.37.16...v29.37.17
@@ -5333,6 +5337,10 @@ patch operations can re-use these classes for generating patch messages.
 [29.37.8]: https://github.com/linkedin/rest.li/compare/v29.37.7...v29.37.8
 [29.37.7]: https://github.com/linkedin/rest.li/compare/v29.37.6...v29.37.7
 [29.37.6]: https://github.com/linkedin/rest.li/compare/v29.37.5...v29.37.6
+=======
+[Unreleased]: https://github.com/linkedin/rest.li/compare/v29.37.19...master
+[29.37.19]: https://github.com/linkedin/rest.li/compare/v>>>>>>> a193ca945 (Pass the UDS address from TransportClientFactory)...v29.37.19
+>>>>>>> a193ca945 (Pass the UDS address from TransportClientFactory)
 [29.37.5]: https://github.com/linkedin/rest.li/compare/v29.37.4...v29.37.5
 [29.37.4]: https://github.com/linkedin/rest.li/compare/v29.37.3...v29.37.4
 [29.37.3]: https://github.com/linkedin/rest.li/compare/v29.37.2...v29.37.3